Struct conway_rs::grid::Grid

pub struct Grid {
    pub cells: Vec<Vec<Cell>>,
    // some fields omitted
}

Grid abstraction as 2D plane containing the cells and its state

Members:

cells

width

height

Methods:

new

new_generation

random_grid

cell

neighbors

Fields
cells: Vec<Vec<Cell>>
Methods

impl Grid

fn new(width: usize, height: usize) -> Grid

Create a new grid based on the width and height.

fn random_grid(width: usize, height: usize) -> Grid

Generate a new grid using random numbers to determine the state of the cells

fn new_generation(self) -> Grid

Create a new generation for all cells applying the rules of Conway's Game of Life (defined on Cell::update) based on the old state and the neighbors.

fn cell(&self, x: usize, y: usize) -> &Cell

access the internal vector of grid based on the 2D point (x,y). Returns the cell.

fn neighbors(&self, cell: &Cell) -> u8

Return the number of neighbors alive from a given cell.